Refactor into EnsureRoles
