+++ lib/cadet_web/views/view_helpers.ex
@@ -8,9 +8,9 @@ defmodule CadetWeb.ViewHelpers do
     conn.assigns[:current_user] != nil
   end
 
-  def is_admin?(conn) do
+  def is_roles?(conn, roles) do
     if logged_in?(conn) do
-      conn.assigns[:current_user].role == :admin
+      conn.assigns[:current_user].role in roles
     else
       false
     end
